Copied the old "player input" to the 2024 project, started adjusting some positions.
I fixed the head area, weapons and all of the numbering for the proficiencies.
Started to work on a 2-column class features field but I am not happy, since "\raggedcolumn" doesn't work.

Still have to work on the lower part and on page two, but left it there for now and copied the "proficiency-markers" file over.

Planning to relocate those markers nex.